LPS-60686 Move everything to internal package in layout set prototype impl bundle
