+++ modules/apps/layout/layout-set-prototype-impl/bnd.bnd
@@ -1,5 +1,5 @@
 Bundle-Name: Liferay Layout Set Prototype Implementation
 Bundle-SymbolicName: com.liferay.layout.set.prototype.impl
 Bundle-Version: 1.0.0
-Export-package: com.liferay.layout.set.prototype.lar
+Export-package: com.liferay.layout.set.prototype.internal.lar
 Include-Resource: classes
