+++ modules/apps/layout/layout-set-prototype-test/src/testIntegration/java/com/liferay/layout/set/prototype/lar/test/LayoutSetPrototypeStagedModelDataHandlerTest.java
@@ -15,7 +15,7 @@
 package com.liferay.layout.set.prototype.lar.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.layout.set.prototype.lar.LayoutSetPrototypeStagedModelDataHandler;
+import com.liferay.layout.set.prototype.internal.lar.LayoutSetPrototypeStagedModelDataHandler;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.TransactionalTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
